+++ Source/ACE.Diagnostics/FrmMonitor.cs
@@ -1,5 +1,4 @@
-﻿using ACE.Common;
-using ACE.Entity;
+﻿using ACE.Entity;
 using ACE.Entity.Enum;
 using System;
 using System.ComponentModel;