+++ Source/ACE/Entity/DebugObject.cs
@@ -8,92 +8,9 @@ namespace ACE.Entity
 {
     public class DebugObject : CollidableObject
     {
-        public DebugObject(ObjectType type, ObjectGuid guid, string name, ushort weenieClassId, ObjectDescriptionFlag descriptionFlag, WeenieHeaderFlag weenieFlag, Position position)
-            : base(type, guid, name, weenieClassId, descriptionFlag, weenieFlag, position)
+        public DebugObject(AceObject aceObject)
+            : base(aceObject)
         {
-        }
-
-        public DebugObject(ObjectGuid guid,  AceObject baseAceObject)
-            : base((ObjectType)baseAceObject.ItemType, guid)
-        {
-            Name = baseAceObject.Name ?? "NULL";
-            DescriptionFlags = (ObjectDescriptionFlag)baseAceObject.AceObjectDescriptionFlags;
-            WeenieClassid = baseAceObject.WeenieClassId;
-            PhysicsData.MTableResourceId = baseAceObject.MotionTableId;
-            PhysicsData.Stable = baseAceObject.SoundTableId;
-            PhysicsData.CSetup = baseAceObject.ModelTableId;
-            PhysicsData.Petable = baseAceObject.PhysicsTableId;
-            PhysicsData.PhysicsState = (PhysicsState)baseAceObject.PhysicsState;
-            PhysicsData.ObjScale = baseAceObject.DefaultScale ?? 0u;
-            PhysicsData.AnimationFrame = baseAceObject.AnimationFrameId;
-            PhysicsData.Translucency = baseAceObject.Translucency ?? 0u;
-            PhysicsData.DefaultScript = baseAceObject.DefaultScript;
-            PhysicsData.DefaultScriptIntensity = (float?)baseAceObject.PhysicsScriptIntensity;
-            PhysicsData.Elasticity = baseAceObject.Elasticity;
-            PhysicsData.Parent = baseAceObject.Parent;
-            PhysicsData.ParentLocation = baseAceObject.ParentLocation;
-            PhysicsData.Friction = baseAceObject.Friction;
-            if (baseAceObject.CurrentMotionState == "0" || baseAceObject.CurrentMotionState == null)
-                PhysicsData.CurrentMotionState = null;
-            else
-                PhysicsData.CurrentMotionState = new UniversalMotion(Convert.FromBase64String(baseAceObject.CurrentMotionState));
-
-            // game data min required flags;
-            Icon = baseAceObject.IconId;
-            AmmoType = (AmmoType?)baseAceObject.AmmoType;
-            Burden = baseAceObject.Burden;
-            CombatUse = (CombatUse?)baseAceObject.CombatUse;
-            ContainerCapacity = baseAceObject.ContainersCapacity;
-            Cooldown = baseAceObject.CooldownId;
-            CooldownDuration = baseAceObject.CooldownDuration;
-            HookItemTypes = baseAceObject.HookItemTypes;
-            HookType = baseAceObject.HookType;
-            IconOverlay = baseAceObject.IconOverlayId;
-            IconUnderlay = baseAceObject.IconUnderlayId;
-            ItemCapacity = baseAceObject.ItemsCapacity;
-            Material = (Material?)baseAceObject.MaterialType;
-            MaxStackSize = baseAceObject.MaxStackSize;
-            Wielder = baseAceObject.WielderId;
-            ContainerId = baseAceObject.ContainerId;
-            ClothingBase = baseAceObject.ClothingBase;
-            CurrentWieldedLocation = (EquipMask?)baseAceObject.CurrentWieldedLocation;
-
-            // TODO: this needs to be pulled in from pcap data. Missing - Name Plural never set need to address
-
-            Priority = (CoverageMask?)baseAceObject.Priority;
-            RadarBehavior = (RadarBehavior?)baseAceObject.Radar;
-            RadarColor = (RadarColor?)baseAceObject.BlipColor;
-            Script = baseAceObject.PhysicsScript;
-            Spell = (Spell?)baseAceObject.SpellId;
-            StackSize = baseAceObject.StackSize;
-            Structure = baseAceObject.Structure;
-            TargetType = baseAceObject.TargetTypeId;
-            GameDataType = baseAceObject.ItemType;
-            UiEffects = (UiEffects?)baseAceObject.UiEffects;
-            Usable = (Usable?)baseAceObject.ItemUseable;
-            UseRadius = baseAceObject.UseRadius;
-            ValidLocations = (EquipMask?)baseAceObject.ValidLocations;
-            Value = baseAceObject.Value;
-            Workmanship = baseAceObject.Workmanship;
-
-            PhysicsData.SetPhysicsDescriptionFlag(this);
-            WeenieFlags = SetWeenieHeaderFlag();
-            WeenieFlags2 = SetWeenieHeaderFlag2();
-
-            baseAceObject.AnimationOverrides.ForEach(ao => ModelData.AddModel(ao.Index, ao.AnimationId));
-            baseAceObject.TextureOverrides.ForEach(to => ModelData.AddTexture(to.Index, to.OldId, to.NewId));
-            baseAceObject.PaletteOverrides.ForEach(po => ModelData.AddPalette(po.SubPaletteId, po.Offset, po.Length));
-            ModelData.PaletteGuid = baseAceObject.PaletteId;
-        }
-
-        public DebugObject(AceObject aceO)
-            : this(new ObjectGuid(aceO.AceObjectId), aceO)
-        {
-            // FIXME(ddevec): These should be inhereted from aceO, not copied
-            Location = aceO.Location;
-            Debug.Assert(aceO.Location != null, "Trying to create DebugObject with null location");
-            WeenieClassid = aceO.WeenieClassId;
-            GameDataType = aceO.ItemType;
         }
 
         public override void OnCollide(ObjectGuid playerId)